feat: add check id on update and delete action
- add check id before action
- add truncate command

+++ api/v1/books/index.ts
@@ -5,7 +5,7 @@ const prisma = new PrismaClient()
 
 export default async function handler(req: VercelRequest, res: VercelResponse) {
     try {
-        let { idStr } = req.query
+        let idStr = req.query.id
         if (Array.isArray(idStr)) {
             idStr = idStr[0]
         }
@@ -17,11 +17,27 @@ export default async function handler(req: VercelRequest, res: VercelResponse) {
                     const book = await prisma.book.findFirst({
                         where: { id }
                     })
-                    return res.status(200).json(book)
+                    if (book === null) {
+                        return res.status(404).json({
+                            status: false,
+                            message: "book not found",
+                            data: null
+                        })
+                    }
+
+                    return res.status(200).json({
+                        status: true,
+                        message: "success get detail book",
+                        data: book
+                    })
                 }
 
                 const books = await prisma.book.findMany()
-                return res.status(200).json(books)
+                return res.status(200).json({
+                    status: true,
+                    message: "success get list of book",
+                    data: books
+                })
             }
 
             case "POST": {
@@ -29,34 +45,88 @@ export default async function handler(req: VercelRequest, res: VercelResponse) {
                 const book = await prisma.book.create({
                     data
                 })
-                return res.status(201).json(book)
+                return res.status(201).json({
+                    status: true,
+                    message: "success create a book",
+                    data: book
+                })
             }
 
             case "PUT": {
+                const b = await prisma.book.findFirst({
+                    select: {
+                        id: true
+                    },
+                    where: { id }
+                })
+                if (b === null) {
+                    return res.status(404).json({
+                        status: false,
+                        message: "book not found",
+                        data: null
+                    })
+                }
+
                 let data = req.body as Prisma.BookUpdateInput
                 const book = await prisma.book.update({
                     data,
                     where: { id }
                 })
-                return res.status(200).json(book)
+                return res.status(200).json({
+                    status: true,
+                    message: "success update a book detail",
+                    data: book
+                })
             }
 
             case "DELETE": {
+                if (req.query.isDelete == "all") {
+                    await prisma.$executeRaw`TRUNCATE TABLE Book;`
+                    return res.status(200).json({
+                        status: true,
+                        message: "success delete all book",
+                        data: null
+                    })
+                }
+
+                const b = await prisma.book.findFirst({
+                    select: {
+                        id: true
+                    },
+                    where: { id }
+                })
+                if (b === null) {
+                    return res.status(404).json({
+                        status: false,
+                        message: "book not found",
+                        data: null
+                    })
+                }
+
                 const book = await prisma.book.delete({
                     where: { id }
                 })
-                return res.status(201).json(book)
+                return res.status(200).json({
+                    status: true,
+                    message: "success delete book",
+                    data: book
+                })
             }
+
             default: {
                 return res.status(404).json({
-                    "message": "Route Not Found"
+                    status: false,
+                    message: "bad route",
+                    data: null
                 })
             }
         }
     } catch(err) {
-        console.error("error: ", err)
+        console.error("catch-a-runtime-error: ", err)
         return res.status(500).json({
-            "message": "Internal Server Error"
+            status: false,
+            message: "internal server error",
+            data: null
         })
     }
 }
